refactor: extract common explore widgets
